APB-6119 adding search/filter to mangage group clients. first stage -> refactor
