+++ test/controllers/ManageGroupControllerSpec.scala
@@ -843,370 +843,6 @@ class ManageGroupControllerSpec extends BaseSpec {
     }
   }
 
-  s"GET ${routes.ManageGroupController.showExistingGroupClients(accessGroup._id.toString)}" should {
-
-    "render correctly the EXISTING CLIENTS page" in {
-      //given
-      val groupWithClients = accessGroup.copy(clients =
-        Some(displayClients.map(DisplayClient.toEnrolment).toSet))
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-
-      expectGetGroupSuccess(accessGroup._id.toString, Some(groupWithClients))
-
-
-      //when
-      val result =
-        controller.showExistingGroupClients(groupWithClients._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title shouldBe "Manage clients - Bananas - Manage Agent Permissions - GOV.UK"
-      html.select(Css.PRE_H1).text shouldBe "Bananas"
-      html.select(Css.H1).text shouldBe "Manage clients"
-
-      val th = html.select(Css.tableWithId("sortable-table")).select("thead th")
-      th.size() shouldBe 3
-      th.get(0).text() shouldBe "Client name"
-      th.get(1).text() shouldBe "Tax reference"
-      th.get(2).text() shouldBe "Tax service"
-
-      val trs = html.select(Css.tableWithId("sortable-table")).select("tbody tr")
-
-      trs.size() shouldBe 3
-      //first row
-      trs.get(0).select("td").get(0).text() shouldBe "friendly0"
-      trs.get(0).select("td").get(1).text() shouldBe "ending in 6780"
-      trs.get(0).select("td").get(2).text() shouldBe "VAT"
-
-      //last row
-      trs.get(2).select("td").get(0).text() shouldBe "friendly2"
-      trs.get(2).select("td").get(1).text() shouldBe "ending in 6782"
-      trs.get(2).select("td").get(2).text() shouldBe "VAT"
-
-      html.select("p#clients-in-group").text() shouldBe "Showing total of 3 clients"
-      html.select("a#update-clients").text() shouldBe "Update clients"
-      html.select("a#update-clients").attr("href") shouldBe
-        routes.ManageGroupController.showManageGroupClients(accessGroup._id.toString).url
-    }
-
-  }
-
-  s"GET ${routes.ManageGroupController.showManageGroupClients(accessGroup._id.toString)}" should {
-
-    "render correctly the manage group CLIENTS page" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-      stubGetClientsOk(arn)(fakeClients)
-
-      //when
-      val result =
-        controller.showManageGroupClients(accessGroup._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title shouldBe "Select clients - Manage Agent Permissions - GOV.UK"
-      html.select(Css.PRE_H1).text shouldBe "Bananas access group"
-      html.select(Css.H1).text shouldBe "Select clients"
-
-      val th = html.select(Css.tableWithId("sortable-table")).select("thead th")
-      th.size() shouldBe 4
-      th.get(1).text() shouldBe "Client name"
-      th.get(2).text() shouldBe "Tax reference"
-      th.get(3).text() shouldBe "Tax service"
-      val trs =
-        html.select(Css.tableWithId("sortable-table")).select("tbody tr")
-
-      trs.size() shouldBe 3
-      //first row
-      trs.get(0).select("td").get(1).text() shouldBe "friendly0"
-      trs.get(0).select("td").get(2).text() shouldBe "ending in 6780"
-      trs.get(0).select("td").get(3).text() shouldBe "VAT"
-
-      //last row
-      trs.get(2).select("td").get(1).text() shouldBe "friendly2"
-      trs.get(2).select("td").get(2).text() shouldBe "ending in 6782"
-      trs.get(2).select("td").get(3).text() shouldBe "VAT"
-    }
-
-    "render correctly the manage group CLIENTS page when there are no clients" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-      stubGetClientsAccepted(arn)
-
-      //when
-      val result =
-        controller.showManageGroupClients(accessGroup._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title shouldBe "Select clients - Manage Agent Permissions - GOV.UK"
-      html.select(Css.PRE_H1).text shouldBe "Bananas access group"
-      html.select(Css.H1).text shouldBe "Select clients"
-
-      val th = html.select(Css.tableWithId("sortable-table")).select("thead th")
-      th.size() shouldBe 4
-      th.get(1).text() shouldBe "Client name"
-      th.get(2).text() shouldBe "Tax reference"
-      th.get(3).text() shouldBe "Tax service"
-      val trs =
-        html.select(Css.tableWithId("sortable-table")).select("tbody tr")
-
-      trs.size() shouldBe 0
-    }
-
-    "render correctly the manage group CLIENTS page when there are clients already in the group" in {
-      //given
-      val enrolments = displayClients.map(dc => DisplayClient.toEnrolment(dc)).toSet
-      val groupWithClients = accessGroup.copy(clients = Some(enrolments))
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-
-      expectGetGroupSuccess(groupWithClients._id.toString, Some(groupWithClients))
-
-      stubGetClientsOk(arn)(fakeClients)
-
-      //when
-      val result =
-        controller.showManageGroupClients(groupWithClients._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title shouldBe "Select clients - Manage Agent Permissions - GOV.UK"
-      html.select(Css.PRE_H1).text shouldBe "Bananas access group"
-      html.select(Css.H1).text shouldBe "Select clients"
-
-      val th = html.select(Css.tableWithId("sortable-table")).select("thead th")
-      th.size() shouldBe 4
-      th.get(1).text() shouldBe "Client name"
-      th.get(2).text() shouldBe "Tax reference"
-      th.get(3).text() shouldBe "Tax service"
-      val trs =
-        html.select(Css.tableWithId("sortable-table")).select("tbody tr")
-
-      trs.size() shouldBe 3
-      //first row
-      trs.get(0).select("td").get(1).text() shouldBe "friendly0"
-      trs.get(0).select("td").get(2).text() shouldBe "ending in 6780"
-      trs.get(0).select("td").get(3).text() shouldBe "VAT"
-
-      //last row
-      trs.get(2).select("td").get(1).text() shouldBe "friendly2"
-      trs.get(2).select("td").get(2).text() shouldBe "ending in 6782"
-      trs.get(2).select("td").get(3).text() shouldBe "VAT"
-    }
-
-    "render with clients held in session when a filter was applied" in {
-
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(
-        sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients.take(1)))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-
-      stubGetClientsOk(arn)(fakeClients)
-
-      val result =
-        controller.showManageGroupClients(accessGroup._id.toString)(request)
-
-      status(result) shouldBe OK
-
-      val html = Jsoup.parse(contentAsString(result))
-
-      html.title() shouldBe "Select clients - Manage Agent Permissions - GOV.UK"
-      html.select(Css.H1).text() shouldBe "Select clients"
-
-      val trs =
-        html.select(Css.tableWithId("sortable-table")).select("tbody tr")
-
-      trs.size() shouldBe 1
-      trs.get(0).select("td").get(1).text() shouldBe "friendly0"
-      trs.get(0).select("td").get(2).text() shouldBe "ending in 6780"
-      trs.get(0).select("td").get(3).text() shouldBe "VAT"
-    }
-  }
-
-  s"POST ${routes.ManageGroupController.submitManageGroupClients(accessGroup._id.toString).url}" should {
-
-    "save selected clients to session" when {
-
-      s"button is Continue and redirect to ${routes.ManageGroupController.showManageGroups.url}" in {
-
-        val encodedDisplayClients = displayClients.map(
-          client =>
-            Base64.getEncoder.encodeToString(
-              Json.toJson(client).toString.getBytes))
-
-        implicit val request =
-          FakeRequest("POST", routes.ManageGroupController.submitManageGroupClients(accessGroup._id.toString).url)
-            .withFormUrlEncodedBody(
-              "hasSelectedClients" -> "false",
-              "clients[0]" -> encodedDisplayClients.head,
-              "clients[1]" -> encodedDisplayClients.last,
-              "search" -> "",
-              "filter" -> "",
-              "continue" -> "continue"
-            )
-            .withSession(SessionKeys.sessionId -> "session-x")
-
-        await(sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients.take(1)))
-        await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-
-        expectAuthorisationGrantsAccess(mockedAuthResponse)
-        expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-
-        expectUpdateGroupSuccess(accessGroup._id.toString,
-          UpdateAccessGroupRequest(clients = Some(Set(displayClients.head, displayClients.last).map(toEnrolment(_)))))
-
-        val result =
-          controller.submitManageGroupClients(accessGroup._id.toString)(request)
-
-        status(result) shouldBe SEE_OTHER
-        redirectLocation(result).get shouldBe
-          routes.ManageGroupController.showReviewSelectedClients(accessGroup._id.toString).url
-        await(sessionCacheRepo.getFromSession(FILTERED_CLIENTS)) shouldBe Option.empty
-        await(sessionCacheRepo.getFromSession(HIDDEN_CLIENTS_EXIST)) shouldBe Option.empty
-      }
-
-      "display error when button is Continue, no filtered clients, no hidden clients exist and no clients were selected" in {
-        // given
-
-        implicit val request = FakeRequest(
-          "POST",
-          routes.ManageGroupController.submitManageGroupClients(accessGroup._id.toString)
-            .url
-        ).withFormUrlEncodedBody(
-            "hasSelectedClients" -> "false",
-            "clients" -> "",
-            "search" -> "",
-            "filter" -> "",
-            "continue" -> "continue"
-          ).withSession(SessionKeys.sessionId -> "session-x")
-
-        expectAuthorisationGrantsAccess(mockedAuthResponse)
-        await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-        expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-        stubGetClientsOk(arn)(fakeClients)
-
-
-        // when
-        val result = controller.submitManageGroupClients(accessGroup._id.toString)(request)
-
-        status(result) shouldBe OK
-        val html = Jsoup.parse(contentAsString(result))
-
-        // then
-        html.title() shouldBe "Error: Select clients - Manage Agent Permissions - GOV.UK"
-        html.select(Css.H1).text() shouldBe "Select clients"
-        html
-          .select(Css.errorSummaryForField("clients"))
-        await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)).isDefined shouldBe false
-      }
-
-      "display error when filtered clients and form has errors" in {
-        // given
-
-        implicit val request = FakeRequest(
-          "POST",
-          routes.ManageGroupController.submitManageGroupClients(accessGroup._id.toString).url
-        ).withFormUrlEncodedBody(
-          "hasSelectedClients" -> "false",
-          "clients" -> "",
-          "search" -> "",
-          "filter" -> "",
-          "submitFilter" -> "submitFilter"
-        ).withSession(SessionKeys.sessionId -> "session-x")
-
-        expectAuthorisationGrantsAccess(mockedAuthResponse)
-        await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-        expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-        await(sessionCacheRepo.putSession(FILTERED_CLIENTS, displayClients))
-
-        // when
-        val result = controller.submitManageGroupClients(accessGroup._id.toString)(request)
-
-        status(result) shouldBe OK
-        val html = Jsoup.parse(contentAsString(result))
-
-        // then
-        html.title() shouldBe "Error: Select clients - Manage Agent Permissions - GOV.UK"
-        html.select(Css.H1).text() shouldBe "Select clients"
-        html
-          .select(Css.errorSummaryForField("clients"))
-        await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)).isDefined shouldBe false
-      }
-
-
-      s"when Filter clicked redirect to ${routes.ManageGroupController.showManageGroupClients(accessGroup._id.toString).url}" in {
-
-        implicit val request = FakeRequest(
-          "POST",
-          routes.ManageGroupController.submitManageGroupClients(accessGroup._id.toString).url
-        ).withFormUrlEncodedBody(
-          "hasSelectedClients" -> "false",
-          "clients" -> "",
-          "search" -> "",
-          "filter" -> "Ab",
-          "submitFilter" -> "submitFilter"
-        ).withSession(SessionKeys.sessionId -> "session-x")
-
-        await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-        expectAuthorisationGrantsAccess(mockedAuthResponse)
-        expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-        stubGetClientsOk(arn)(fakeClients)
-
-        // when
-        val result = controller.submitManageGroupClients(accessGroup._id.toString)(request)
-        status(result) shouldBe SEE_OTHER
-      }
-    }
-  }
-
-  s"GET ${routes.ManageGroupController.showReviewSelectedClients(accessGroup._id.toString)}" should {
-
-
-    "redirect if no clients selected are in session" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-
-      //when
-      val result = controller.showReviewSelectedClients(accessGroup._id.toString)(request)
-
-      //then
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result).get shouldBe routes.ManageGroupController.showManageGroupClients(accessGroup._id.toString).url
-    }
-
-    "render correctly the manage group REVIEW SELECTED page" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-
-      //when
-      val result = controller.showReviewSelectedClients(accessGroup._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-      val html = Jsoup.parse(contentAsString(result))
-      html.title() shouldBe "Review selected clients - Manage Agent Permissions - GOV.UK"
-      html.select(H1).text() shouldBe "You have selected 3 clients"
-      html.select(Css.tableWithId("sortable-table")).select("tbody tr").size() shouldBe 3
-
-    }
-  }
-
   s"POST ${routes.ManageGroupController.submitAddUnassignedClients}" should {
     s"save selected unassigned clients and redirect to ${routes.ManageGroupController.showSelectedUnassignedClients} " +
       s"when button is Continue" in {
@@ -1316,42 +952,6 @@ class ManageGroupControllerSpec extends BaseSpec {
 
       await(sessionCacheRepo.getFromSession(SELECTED_CLIENTS)) shouldBe None
 
-    }
-  }
-
-  s"GET ${routes.ManageGroupController.showGroupClientsUpdatedConfirmation(accessGroup._id.toString).url}" should {
-
-    "render correctly" in {
-      //given
-      await(sessionCacheRepo.putSession(OPTIN_STATUS, OptedInReady))
-      await(sessionCacheRepo.putSession(SELECTED_CLIENTS, displayClients))
-      expectAuthorisationGrantsAccess(mockedAuthResponse)
-      expectGetGroupSuccess(accessGroup._id.toString, Some(accessGroup))
-
-      //when
-      val result = controller.showGroupClientsUpdatedConfirmation(accessGroup._id.toString)(request)
-
-      //then
-      status(result) shouldBe OK
-
-      //and
-      val html = Jsoup.parse(contentAsString(result))
-      html.title shouldBe "Bananas access group clients updated - Manage Agent Permissions - GOV.UK"
-      html
-        .select(Css.confirmationPanelH1)
-        .text() shouldBe "Bananas access group clients updated"
-      html.select(Css.H2).text() shouldBe "What happens next"
-      html
-        .select(Css.paragraphs)
-        .get(0)
-        .text() shouldBe "You have changed the clients that can be managed by the team members in this access group."
-      html
-        .select("a#returnToDashboard")
-        .text() shouldBe "Return to manage access groups"
-      html
-        .select("a#returnToDashboard")
-        .attr("href") shouldBe routes.ManageGroupController.showManageGroups.url
-      html.select(Css.backLink).size() shouldBe 0
     }
   }
